S4/AWS/requirements.txt
https://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp38-cp38-linux_x86_64.whl
requests_toolbelt
torchvision==0.6.0
facenet-pytorch
numpy==1.18.5


